requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
tiktoken
openai
